pyproject.toml
[project]
name = "globeco-portfolio-service"
version = "0.1.0"
description = "The GlobeCo Suite Portfolio Service.  GlobeCo is a suite of applications for benchmarking Kubernetes autoscaling."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=1.29.0",
    "dnspython>=2.7.0",
    "fastapi[standard]>=0.115.12",
    "gunicorn>=23.0.0",
    "mongo-migrate>=0.1.2",
    "pydantic-settings>=2.9.1",
    "pytest-asyncio>=0.26.0",
    "pytest>=8.3.5",
    "testcontainers[mongodb]>=4.10.0",
    "opentelemetry-api>=1.34.0",
    "opentelemetry-sdk>=1.34.0",
    "opentelemetry-instrumentation>=0.55b1",
    "opentelemetry-exporter-otlp>=1.34.0",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
    "opentelemetry-instrumentation-grpc>=0.55b1",
    "opentelemetry-instrumentation-logging>=0.55b1",
    "opentelemetry-instrumentation-requests>=0.55b1",
    "opentelemetry-instrumentation-httpx>=0.55b1",
    "prometheus_client>=0.22.0",
]

[tool.setuptools]
packages = ["app"]
